modernize this engine with a Service for updating, encapsulate the http requests in a DataContainer subclass, use KIO instead of QHttp and generally clean stuff up.
note that i haven't had a chance to actually test the engine; it compiles and the logic should be generally correct, but i need to update the applet as well as it utterly breaks with these changes. i also want to add a Service UI to plasmaengineexplorer.

but that's all tomorrow's work =) for today ... we have this.


svn path=/trunk/KDE/kdeplasma-addons/; revision=856493