Pastebin is a PopupApplet now
Now it behaves correctly when it's placed on a panel.
Some more work needs to be done in the UI and UX.
Possibly create a 'history' of pasted URL's.

But this features requests (new UI, history of links, etc..)
should be put inside a new bug. Waiting for pinheiro's icon
and mockup ;)

BUG:183125
CCMAIL:nuno@oxygen-icons.org

svn path=/trunk/KDE/kdeplasma-addons/; revision=931202