Control+Scroll changes font size
One of the things that have been annoying me for a while with the notes
applet is that a common operation (changing the font size) needs going
through the config dialog. I've borrowed a metaphore from the
webbrowser, and now pressing the control key and scrolling up and down
changes the font size displayed. This works surprisingly well since you
get very short feedback cycles between setting fontsize and seeing the
result. It's now very easy to make the contents of the applet exactly
fit the size of the note.

svn path=/trunk/KDE/kdeplasma-addons/; revision=945714