Creating ToolTip for Dict plasmoid
It just contains the name of the applet (calling applet->name() to
avoid new strings due string freeze - and come on, it makes sense ;) ).

I'm not putting the installed sources on the tooltip because it can get
ugly and a mess if someday we have too many sources installed.

Also using init() to handle this creation and other stuff that was being
handled inside constructor.

BUG:183749

svn path=/trunk/KDE/kdeplasma-addons/; revision=970708