TinyURL support on pastebin
Now if you drop an http url on pastebin it will give you back a
tinyurl. Right now I didn't implement this as a backend as we do
not have others for tinyurl, but we may want this to be a backend
for the sake of cleaness.

CCMAIL:plasma-devel@kde.org

svn path=/trunk/KDE/kdeplasma-addons/; revision=993753