Try a nicer icon if can't find a proper one
Just hide the icon if there is no way to get a nice icon.
We already have "today's weather icon" showing on the grid
so use that as it's much nicer and polished.

Hide the icon otherwise and avoid showing dummy icon that
covers the name of the city.

Polish++ :)

svn path=/trunk/KDE/kdeplasma-addons/; revision=1100820