Change weather behavior when it doesnt find icon condition
The correct fix would be to show the bigger version of the current
day condition. It's coming after this one.

This is just polishing as showing an invalid icon in front of
the name of the city is just not good: ugly icon in the face of
the user and broken usability (hard to read the name of the city).
Polishing++

Backport to 4.4


svn path=/branches/KDE/4.4/kdeplasma-addons/; revision=1101199