When pasting raw pixmap data, store the temp file as a PNG, not a JPEG.
When raw image data is dragged to the Pastebin plasmoid, it must save
the data to disk in a temporary file before uploading. Previously, it
used JPEG, but because it's a lossy format, it often looked quite bad,
especially in screenshots including text. Since dragging screenshots
directly from KSnapshot seems to be a very common use case, this was
obviously a (minor) issue for a lot of people. PNG files will yield
much higher quality and often (especially in the simple screenshot
case) smaller uploads.

I would also like to point out that this is the smallest patch I've
ever written. It deletes a 'J' and converts an 'E' to an 'N'. ;)

svn path=/trunk/KDE/kdeplasma-addons/; revision=1124166