New updated pastebin config ui
Remove some fields that are not used anymore (this applet will not support)
private "shares" for now. This is something that I'm still thinking about
if it worth it. If somebody wants a backend that does private stuff, it's
easy enough to write one and we don't need to clutter the UI that should be
as simple as possible for this widget.

svn path=/trunk/KDE/kdeplasma-addons/; revision=1168974