First moves towards the (correct) use of the 'share' dataengine
Slowly start to move the code to use the new dataengine and make the
'pastebin' one deprecated (to be removed soon).
These are the first steps, however we still need to review the UI and
the code paths for pasting stuff.

Now it handles KIO urls (the dataengine itself has this feature) and
the dataengine includes support for some services that were asked as
wishes on bko. Later we can include more with GHNS.

TinyURL support was dropped as it was not much used and a plugin can
be created for that later if a user wishes this kind of service.

FEATURE: Use new 'share' dataengine with support for plugins
BUG: 226397
BUG: 230324
BUG: 227557


svn path=/trunk/KDE/kdeplasma-addons/; revision=1168990