Properly parse the result of the geolocation engine
Just truncate on the "," if there is one. Most of the time the
engine will not return a result with "," so when truncating it
was always empty and we were never trying to auto guess the location
to search weather for.

This fixes this BUG and restores the proper behavior.

svn path=/trunk/KDE/kdeplasma-addons/; revision=1169289