Add support for GHNS on pastebin plasmoid
This way it's easier to add/remove support for different backends
and people can start using opendesktop.org to share their files.

It still has a bug: it's not properly removing the packages. Not sure
where is the problem (maybe in the package implementation on the dataengine?)

Thanks Marco for having a look at this!

CCMAIL:notmart@gmail.com

svn path=/trunk/KDE/kdeplasma-addons/; revision=1190786