Properly compute the height for the Component, Item, Text and Highlight
properly. This is an interesting idea...using an invisible string which
is set to "Arbitrary String", and do property int itemHeight:
itemMetric.height. We'll see how that goes. It works nicely though. My
main goal was to make it so that it's not *too* dynamic, as well as
being able to set in one place without touching 5 others.