Start at full OAuth support
Introducing a new class that does the interaction with twitter and
identi.ca using full OAuth, not xauth-based authentication. Also disable
basic authentication for identi.ca for now, as it's supposed to also
allow full oauth.

This implements the request token step and authorized the app using a
webview and some injected javascript magic, so the user doesn't need to
click herself. :)

Still needs a lot of work, but promising so far. The first two steps
work for me in twitter now. identi.ca not yet.