DisableByDefault duckduckgo, youtube and bing runners
Those runners are sending everything we type on KRunner to external
sites we don't have control on, and we don't want that.

They seem to be workign well, but still we can't enable something by
default it is sending everything without any kind of control.

Please, enable them again when prefixes are working.