Rewire timeline source updating
We have one authhelper per service, each timeline holds a pointer to
check if the authhelper isAuthorized() and starts (or delays starting)
the job when it's authorized.

Right now, we do not store the access token, so each time the engine is
started, we need to go through the whole oauth procedure (and we thus
need a password). Needs fixing.