Don't set current user to requested user info
This should be a bit nicer, since right now, if we request the Timeline
of a user, that user is made the current one, which is wrong if we want
to see someone else's timeline, for example.

This needs deeper going corrections, and such a change would mean
changing assumption currently made by plasmoids. Let's see...