Bootstrapping the weatherstation qml port
The reference to the LCD widget was removed from the WeatherStation
class. My idea is to expose the LCD widget as a QML element and deal
with it in the qml world. It was not done yet because some compilation
problems when the QtDeclarative module is included to use the
qmlRegisterType() function.

Other changes:
    - Created the package structure and the main.qml file
    - CMakeLists.txt updated

Signed-off-by: Luís Gabriel Lima <lampih@gmail.com>