rm bing and duckduckgo runner the latter was an accidental merge (since i did a branch of a branch of a branch and then merged it..so that snuck in).
the bing runner needs to be removed because microsoft very recently
followed suit with all of the other search engines in the world and
decided to close their api and make it paid-only. so basically the only
hope for an image search runner, ever...would be a flickr one (which i
plan). had this been left unchanged, shortly after the 4.9 release the
apis would just stop working.

BUG: 301657
FIXED-IN: 4.9