Add a (hidden) config to manually set media apps.
This is useful for applications that provide an mpris-interface
and are not in the categories AudioVideo or Music.
An example would be a webbrowser, that provides an mpris-interface
for html5-audio and -video. For rekonq there's a pending review
request for exactly that. I will add rekonq to the default config,
as soon as the patch is accepted.

REVIEW: 105137