Fix the plasma spellchecker's 'foreign language' support. Previously this caused segfaults (even if not used) because it called setLanguage(), which is not thread-safe, in match().
Instead, this patch constructs a new speller safely for each new
language, without deleting the old one. Old spellers are instead
deleted on the teardown() signal.

While we're at it, amend the language detection so that the user
can type natural language names (eg, 'german') and have the
spell-checker find the right language.

REVIEW: 106244
BUG: 303831
BUG: 264779
FIXED-IN: 4.9.3