correctly determine input in calculator
The bug reported this behaviour: if you press something like 'a+b+c=',
the result would be incorrect because the calculator plasmoid wrongly
resets the calculator when pressing 'c'. The intended behaviour is that
the calculation is cleared when pressing a number after pressing '='.

This patch adds a variable to determine this situation correctly.

Patch by Paul Rohrbach

CCMAIL:p.b.r@gmx.net
BUG:313869
REVIEW:112674