Remove 'X-Plasma-DefaultSize' declarations in metadata.desktop files
Removed all 'X-Plasma-DefaultSize' declarations in metadata.desktop files where the corrosponding main QML file contained a minimumHeight and minimumWidth declaration.

Below is a listing of the notes I kept for the items I changed/didn't change.

/applets/bubblemon/plasma-applet-bubblemon.desktop:
Ignored as it has not been moved to qml/is not a metadata.desktop file.

/applets/calculator/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/comic/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/konqprofiles/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/konsoleprofiles/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/nowplaying/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/weather/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/weatherstation/package/metadata.desktop:
minimumWidth and mimimumHeight are set in qml, declaration removed.

/applets/webslice/plasma/plasma-applet-webslice.desktop:
Ignored as it has not been moved to qml/is not a metadata.desktop file.

Acked by David Edmundson, Sebas Kügler, Heena Mahour

REVIEW: 114682