Add Disk Quota Plasmoid
The disk quota is usually used in enterprise installations
where network shares are mounted locally. Typically,
sysadmins want to avoid that users copy lots of data into
their folders, and therefor set quotas (the quota limit has
nothing to do with the physical size of a partition).
Typically, once a user gets over the hard limit of the quota,
the account is blocked and the user cannot login anymore.
This happens from time to time, since the users are not
really aware of the current quota limit
and the already used disk space.

Here is where the "Disk Quota" plasmoid helps:
It continusouly monitors the disk quota and
warns the quota apprpriately.

A detailed description including screenshots can be found
in this blog: http://kate-editor.org/?p=3591

patch by Dominik Haumann

REVIEW:124589
CCMAIL:dhaumann@kde.org