comic engine: fix checking of main script extension
When iterating through all the available extensions to check for the
actual extension of the main script, do not replace "mainscript" with
the path of the current iteration step (which has the extension):
doing that causes "mainscript" to get all the available extensions
appended to it.

This bug could be seen when having other kross engines than krossqts
(provided with kross itself), e.g. krosspython.

Regression introduced while porting to frameworks, i.e.
commit 7c37be6fb67806e687f6691f2e74e802d9ff3f0f.