[Color Picker] Fix grabbing color in multi screen environment
I thought I was smart by looking for the proper screen and then calculating the
mouse offset relative to it and so on when in fact the QPixmap created by
QScreen::grabWindow(0) contains all screens, not just the one I grab from.
This is probably once again X weirdness.

BUG: 356464
FIXED-IN: 5.5.1