[showdesktop] Fix sizing of the plasmoid in the panel
Since 5.6, if you add a "Show Desktop" plasmoid to the panel, the
panel's layout gets completely broken (apparently the plasmoid is sized
to infinity).
This sets the minimum size to sane values which fixes the problem.

(Note that this is for the 5.6 branch only, the problem doesn't exist in
 master)

FIXED-IN: 5.6.1
REVIEW: 127499