Fix null values in currency conversion
Summary:
We would check if the string (unit1) matched any known units (s) using a
case insensitve compare.

If it did, we would treat that as a valid unit and pass it on to
KUnitConversion, but unfortunately that is case senstive, and then fails
to find anything.

This fixes the case of the user typed unit to be correct.

Test Plan:
Ran krunner with "640 gbp in eur" it now returns the same as "640 GBP
in EUR" .
debug in kunitconversion shows it now has a valid unit ID for the first
query.
(the result is still utterly wrong, but it's one problem less)

Reviewers: #plasma

Subscribers: plasma-devel

Projects: #plasma

Differential Revision: https://phabricator.kde.org/D1338