Correctly translate combo box in media frame plasmoid
Summary:
ListElement doesn't support script values. The original code used a
"stupid hack" to try and run i18n later, but that doesn't work as we
haven't extracted the original text.

Porting to a Javascript map works (and is the approach taken by the main
image wallpaper for a similar model)

Introduces a lot of new strings being extracted properly.

Subscribers: plasma-devel

Projects: #plasma

CC: kde-i18n-docs@kde.org

Differential Revision: https://phabricator.kde.org/D1352