Make media frame start with random image if set to random
Summary:
On startup the media frame will always start with the same image, even if random is selected.

The reason is that the images are added to the list of items one by one and if no item is currently showing one will be picked as soon as it's available. In that case, always the first one that's discovered.

The fix skips to the next (now random) item after loading is done when set to random.

Test Plan:

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D2839