[applets] Support KWin/Wayland in ColorPicker
Summary:
KWin provides a DBus interface for interactive color picking. This
allows to have a color picker on Wayland.

To use this the GrabWidget is refactored and an abstract Grabber class
is split out. The current functionality using an (X11) window is moved
into an X11Grabber class and a new class KWinWaylandGrabber is added
which performs the interaction through DBus.

Reviewers: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3481