Avoid showing wrong timezone suffix in krunner time
Summary:
Krunner is suffixing the local timezone when asked for the time of other regions. QLocale doesn't provide a way to set the locale based only on the country (which is provided by QDateTime), so it would be better if we just use the short time format and avoid showing the timezone altogether.

  - Current time in my zone
{F3745768}

  - Before Patch
{F3745775}

  - After Patch
{F3745776}

Reviewers: davidedmundson, broulik

Reviewed By: davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5770