Replace XML parsing in National Geographic dataengine
The current parsing mechanism to obtain the current NatGeo POTD based on
XML is broken, this patch proposes an alternative mechanism using
QRegularExpression library. For this method to continue working in the
future it is supposed that the web page will maintain a sane format,
namely that the relevant tag will not be broken among many lines, the
fields inside the tag will maintain the same relative order, and
obviously that it will not be obfuscated.

BUG: 379003

Reviewers: sebas, mart

Reviewed By: mart

Subscribers: mart, Zren, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5729