[weather] Add configuration option which weather services providers to use
Summary:
When searching a weather station/location to select for the weather applet,
so far the user had no chance to control which weather services providers
are queried. Instead simply all are given the search string. Which can both
result in unwanted hits due to regions covered, as well as some minimal
privacy breach.

This patch adds a dropdown menu to the search form where the user can
control which services providers are used. The settings selected is stored
in the config for the given applet instance.

The setting defaults to an empty list, so the user has to opt-in to the use
of any provider.

Ideally the dropdown menu listing the providers with checkboxes would stay
open after toggling a selection, but QtComponents Menu seems to not allow
any modification of that behaviour. This is a small annoyance, but only once
in a while.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9751