Allow users to change the format to BCD (decimal) in Binary Clock widget
Hours, minutes and seconds are in hexadecimal system. This is sexagesimal format according to Wikipedia. In Plasma4 the format was BCD (Binary-coded decimal) which means that each digit was encoded separately in decimal system.

My fix allows users to choose the format in config. By default the sexagesimal format is used since that was the format after porting the widget to Plasma5.

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10434