[potd dataengine] Enable development of 3rd-party potd provider plugins
Summary:
Installing the headers of libplasmapotdprovidercore in the include namespace
	plasma/potdprovider
as well as some CMake Config files of the name
	PlasmaPotdProvider
allows it 3rd-party to develop own potd provider plugins without having
to mess with the complete kdeplasma-addons repo.

A potd-provider plugin template in kapptemplate format should speed up
development of new plugins some more.

Test Plan:
Generated a new plugin from the template, built and installed it, selected
it in the POTD wallpaper and got the KDE logo as fetched from techbase
shown.

Reviewers: xuetianweng

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11964