Use consistent spacing and units for suffixes in spinboxes
Summary:
As specified by the International System of Units (SI) there is usually a
space between the number and the unit (symbol). To still allow localisations
to apply some custom appendix of the unit, the separating space is part of the
string to translate.

For common units like seconds and minutes the SI symbols "s" & "min" are
used, while for others more explicit terms are used ("days", "strips per
comic").

Reviewers: #plasma, #localization, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: aacid, davidedmundson, plasma-devel

Tags: #plasma, #localization

Differential Revision: https://phabricator.kde.org/D12137