Sync systemloadviewer showed items to ksysguard/system monitor
Summary:
We currently show how much cache is used. It's a completely useless value to know vs how much memory is free and leaves this applet out of sync with all the other memory monitors.

It's problematic because the bar is inconsistent with the tooltip, and
by default the system load viewer chooses a colour very similar to the
application used memory making the UI actually worse.

BUG: 391918

Test Plan:
Ran it
(there's some unrelated bugs/warnings in this applet that need fixing)
Bar showed a sensible value that matched tooltip, rather than
looking 100% full all the time

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: rkflx, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D11558