[weather applet] Make wind direction arrows more breezish
Following status panel type icon design: 22x22 px with 3 px margin

Design considerations:
The same arrow shape should be always shown, only rotated to
point in the respective direction. When rotated in a non-90 ° angle,
any ends should not overlap into the margins of the square bounding box.
To ensure that, the original arrow at 0 ° rotation is done to fit into
a circle bound into the icon content area.
The arrows need to point also in angles 22.5 ° and 67.5 °
for the *** directions. For which there is no equivalent ratio of
integers/pixels, which then could be reused for the arrow shape to ensure
similarly aliased-looking arrow borders. Thus this has not been taken
into account.
The arrow itself follows a shape often used for wind direction arrows.
Details like a container of the windrose, windrose cross or center of
the needle have been left out, as they are not needed given the context.

First approach, might need to see more iterations.