[Timer applet] Fix double speed countdown & commands run multiple times
Summary:
Both the compact and the fullpresentation variants of the applet had
an instance of a timer, and both were connected to the central "running"
flag. So if both variants had been created, there were 2 timers triggering
the decreasing of the seconds, and also firing off the command once done.

Moving the countdown timer (and at the same time also a timer for
delayed saving of the state, which had the same issue) to the central root
item fixes this.

BUG: 381173
FIXED-IN: 5.12.6

Test Plan:
Applet still works as before on panel and background pane, now does proper
countdown in both cases.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D13065