[Timer applet] Reset opacity on timer start and reset
Summary:
When the Timer plasmoid is paused an opacity animation is used to
signal this state. After resuming, restarting or resetting the timer
the current opacity value is not reset which leads to nearly
invisible and hard to read timer digits sometimes.
This patch sets the opacity to 100% on timer (re)start and reset.

BUG: 396497

Test Plan:
* Start Timer plasmoid with e.g. 5 minutes countdown
* Pause Timer by click
* When opacity is low
  * Resume by click or
  * Restart with predefined time or
  * Reset timer
* Opacity of running timer should always reset to 100%

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14701