[Notes] Fix icons being almost invisible with light backgrounds when using dark themes
Summary:
Old code had no way to adapt icon colors based on background so this patch uses different components to color the icons black with light backgrounds, and white with dark backgrounds.

Because Plasma Components 2.0 ToolButton doesn't support icon colors and because with 3.0 the code doesn't work, we have to port the affected buttons to QQC2 ToolButton.

BUG: 353819
FIXED-IN: 5.17

Test Plan:
Before:
{F6808620}

After:
{F6808621}

NOTE: The buttons stay highlighted after being pressed, but I'm trying to solve that in D21026

Reviewers: #plasma, #vdg, ngraham, davidedmundson

Reviewed By: #plasma, #vdg, ngraham, davidedmundson

Subscribers: broulik, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21027