Remove colon (:) prefix when looking up dictionary word
The dictionary engine lets us specify a server and dictionary but we don't need any of that,
and the engine can deal with just a word (these days?).
More importantly, the key inserted into the m_lockers map is without a colon and since sourceRemoved is also called without it,
the wait condition is never met and we time out eventually.
The colon has been in there forever but from current dictionary engine code I don't see why it would be neccessary.

CHANGELOG: Fixed dictionary runner not finding any definitions

BUG: 376905
FIXED-IN: 5.16.4

Differential Revision: https://phabricator.kde.org/D22814