In the note applet, always paste without formatting by default. Add an action to paste preserving formatting in the contextual menu.
Summary:
It frequently happens to me that formatting is preserved when pasting into the note, making the note a bit confusing. I added a button to remove formatting from the pasted text so that I can apply a different one with the other buttons. For example this is what I get when pasting from konsole:

{F7104991}

but I'd personally prefer to be able to do something like this, in this specific case:

{F7104992}

Reviewers: davidedmundson, #plasma

Reviewed By: davidedmundson, #plasma

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22774