Correctly highlight the icon in the system tray
Summary:
Icon is highlighted when CapsLock or NumLock are locked. It should not
be used this way, icon should be highlithed on mouse hover, as any other
icon in system tray.

CCBUG: 400295

Test Plan: Hover mouse over the icon - highlight only when mouse is over the icon. It should not highlight when CapsLock or NumLock are used.

Reviewers: #plasma, davidedmundson, apol, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24737