[applets/Weather] Port config UI mostly to QQC2 and away from Plasma components
Summary:
Everything is ported to QQC2 except for the TableView since there is no direct analogue
in QQC2. The correct porting path is a ScrollView+ListView, but I can't do this right
now since the search/weather station selection functionality is currently broken for me,
making it impossible to test changes to the table.

Test Plan: Everything that was working before still works now

Reviewers: spstarr, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25509