[applets/notes] Rename to "Sticky Note"
Summary:
The current name "Notes" is ambiguous and grammatically incorrect in various places due
to being a plural noun rather than a singular noun.

For example the string "Notes settings" is grammatically incorrect, and "Remove notes"
is ambiguous as to which notes exactly will be removed.

Just changing it to "Note" would fix the grammatical issue but preserve the ambiguity
in a different way though. So this patch changes the name to "Sticky Note" which should
be clear and grammatically correct everywhere.

Test Plan:
Before: {F8116212}

After: {F8116211}

Reviewers: #plasma, #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27532