[applets/notes] Actually hide button row when it's not visible
Summary:
Currently the row's opacity is changed from 0% to 100% depending on focus state, but it
does not actually become hidden when at 0% opacity. This means that it's possible to
click the invisible buttons accidentally when focusing the applet. This has been an
issue for a long time, but becomes dangerous with D28064, since you could delete a note
just by clicking in the bottom-right corner to focus the applet!

This patch makes the button row actually hidden when the applet is not focused, not just
100% transparent.

Test Plan:
Clicking on the place where a button will appear on an unfocused applet now just focuses
it and does not also click the button.

Reviewers: #plasma, #vdg, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28209