[applets/weather] Improve UX for selecting a weather station
Summary:
This patch improves the UX of selecting a weather station by re-implementing it a column
of checkboxes that are visible by default at the top of the window, so that selecting a
weather station is follows the local top-to-bottom flow of the page.

When no weather stations are selected (as it is by default), everything else in the
window is disabled so it's completely obvious what you need to do to proceed.

Test Plan: {F8256457}

Reviewers: #vdg, #plasma, broulik, kossebau, spstarr

Reviewed By: #plasma, spstarr

Subscribers: ndavis, spstarr, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29144