move pot kded to kconfigwatcher
it's easier on the eyes and also works properly... because of how
configs
can cascade across multiple directories the previous fixed path lookup
wasn't all that correct. e.g. initially no file would be found at all,
and
when found it may not actually be the one in the user's home but e.g.
in /etc/xdg