Fix weather widget crash when going to Units tab for the second time
The AbstractUnitListModel is destroyed everytime we leave Units tab
and when we are going to the Units tab for the second time, qml will
try to access AbstractUnitListModel that is already destroyed.
This causes segmentation fault.

Set AbstractUnitListModel parent to qml engine so that the
AbstractUnitListModel is destroyed when qml engine is
destroyed.

BUG: 419709


(cherry picked from commit d50d47ea4be689c46da41a0fa1a2d7955acbcaf7)