Copy result by default
This makes it consistent with what the other runners.
Also it is confusing to have two actions with the same icon.

A similar patch was done to the spellchecker some time ago.

Test Plan:
Running `1m` copies the number to the clipboard

Running `1m` by clicking the action or pressing Shift+Return
copies the number and the resulting unit to the clipboard

(cherry picked from commit 026e84ecbe160993d640beac112e4bcf4257acfa)