[note] Add confirmation dialog when deleting note with real content
We've gotten feedback that it's too easy to accidentally delete a note
now. While there is an undo notification, this is still a fairly scary
thing to have happen if your accidentally-deleted note had important
content in it. Also, the notification is persistent, but only persists
for 60 seconds (plasmashell manually revokes it after that). So if you
accidentally delete a note and go to get coffee or fail to notice the
notification, the content will be lost forever.

To improve the situation here, we now show a warning dialog when you
attempt to delete a note that has user-generated content in it. If the
note is empty, or if the note's content is identical to the top
clipboard entry (as in  he case where the user created the note via
middle-click paste), then it is instead deleted immediately.

Hopefully this will be an adequate compromise between the desire to make
it easy to delete accidentally-created notes and the importance of
keeping user data safe.