[applets/weather] Don't switch focus away from search field to result list
Switching focus away from search field to result list while the user is typing
results in a very annoying search experience.

This change doesn't slow down the user, since the first result is still selected
but not focused. So when user can still directly click "Select" without having
the select the first item.

To make this even nicer, the up and down arrow keys can now be used to navigate
the results list while the search field is still focused.

BUG: 437959
FIXED-IN: 5.23