applets/weather: Improve first-run (needs config) experience
Before, when you first added the weather applet on the panel or the System
Tray, it was completely blank. The only thing to hint that some configuration
was needed was the tooltip saying "Please configure".

With this change, the applet shows some helpful text and a convenient
"Configure" button. It no longer looks broken, and makes it easier for users
to set it up.