Clean up KPackage related code & CMake linking
Plasma::Package is completely deprecated, consequently we should not
rely on it's headers. Also, we do not need to save the package structure
in a static variable, because KPackage caches the instances internally.

With those changes, we do not need to link to plasma-frameworks for the
core engine lib and the package structure.