Fix DictEngine's handling of some errors
Some errors sent by the dictionary server could cause DictEngine to go into an infinite loop.
This adds checks for the ones relevant to the DEFINE and SHOW DB commands.

To me it seems like in cases where it used to freeze it now sometimes slow down for a bit but I wasn't sure how to debug this so I think I am going need some help.