Remove unneeded check for comic Dataengine being valid
BUG: 452596

We do not really care, if the metadata of the plugin is valid, as long as we can
properly create the dataengine from it.

Test Plan:
With plasma-frameworks master, and this on Plasma/5.24 based revision, the dataengine can properly be loaded.

When intenitionally exporting a defunct plugin:
```cpp
class Dummy : public QObject
{
    Q_OBJECT

public:
    Dummy(QObject *parent, const QVariantList &args) {
    }
};
K_PLUGIN_CLASS_WITH_JSON(Dummy, "plasma-dataengine-comic.json")
```
The applet does not crash.