applets/weather: Make desktop widget big enough to fit Configure button
If we don't set Layout.minimum{width,height} properties on the
FullRepresentation, it will default to something tiny and the
Configure… button shown on first launch will overflow.

Set it to a sane size, and also take into account the final layout once
content later gets added so it isn't too small once configured.

BUG: 419591
FIXED-IN: 5.24.7