Remove ContainmentType from the grouping container plasmoid
The grouping container plasmoid is *not* a containment, but is a normal
applet that does some hack to display the contents of the internal
grouping applet which *is* one.

Since plasma-framework is built with
if KCOREADDONS_BUILD_DEPRECATED_SINCE(5, 89)

now everything that has the containmentype key is considered as a
containment, and now the grouping plasmoid is listed in the "add panel"
menu, however it doesn't really work there.
that key was wrong from there to begin with.

BUG: 454416
FIXED-IN: 5.26


(cherry picked from commit 30330a86bb9ade7e468780a3241a95636509142a)