applets/weather: Fix system tray detection
The applet needs to check weather it is in the system tray
to avoid showing a text label beside the icon.

As the property is a flag, it needs a bitwise operator
instead of a plain equality. This also adds a second check
(in practice both are true, but better safe than sorry)

Additionally, also hide the option completely instead of
showing it disabled, which can be confusing.

(cherry picked from commit 271f900e12e577a7bd87a5c0bdaeb6abb9d1be97)