applets/weather: Improve access to the location picker
The current settings page to configure a weather station is mostly
empty space, but it still launches an external dialog to pick a
new location.

We can instead, have the station picker directly in the config page,
so the workflow to select a new location is much more straighforward.

Now that the picker is more prominent, we can also polish some details:
- Add placeholder texts when there is no search in progress
- Port the placeholder to Kiriami.PlaceholderMessage component
- Make the busy indicator more responsive to user typing
- Clear the list when the search field gets empty
- Separate Location and Provider labels: it prevents long strings
  that overflow and force the form layout into narrow/"mobile" mode